+++ config/nightwatch.conf.js
@@ -5,7 +5,6 @@ const print = require('yyl-print');
 const extOs = require('yyl-os');
 const extFs = require('yyl-fs');
 const chalk = require('chalk');
-const seleniumServer = require('selenium-server');
 const chromedriver = require('chromedriver');
 const fs = require('fs');
 const path = require('path');
@@ -13,8 +12,7 @@ const path = require('path');
 const iEnv = util.envParse(process.argv.splice(2).join(' '));
 
 
-const PORT = iEnv.port || 7000;
-const PRODUCT_PORT = PORT + 1;
+const PORT = iEnv.port || 9515;
 
 let USER_CONFIG_PATH = path.join(iEnv.path, 'yyt.config.js');
 
@@ -74,26 +72,16 @@ const DEFAULT_CONFIG = {
     timeout: 30000,
     retry_attempts: 5
   },
-  selenium: {
-    start_process: true,
-    server_path: seleniumServer.path,
-    log_path: false,
-    port: PORT,
-    cli_args: {
-      'webdriver.chrome.driver': chromedriver.path,
-      'webdriver.gecko.driver' : '',
-      'webdriver.edge.driver' : ''
-    }
+  webdriver : {
+    start_process: true
   },
   test_settings: {
     default: {
-      selenium_port: PORT,
-      selenium_host: 'localhost',
-      screenshots: {
-        enabled: false,
-        path: ''
+      webdriver: {
+        server_path: chromedriver.path,
+        port: PORT, // 经测试，修改无效 还是会在 9515 打开
+        cli_args: []
       },
-      silent: true,
       desiredCapabilities: {
         browserName: 'chrome',
         marionette: true,
@@ -101,24 +89,7 @@ const DEFAULT_CONFIG = {
           args: [
           ]
         }
-      },
-      globals: {
-        productListUrl: `http://localhost:${PRODUCT_PORT}/productlist.html`
       }
-    },
-    chrome: {
-      desiredCapabilities: {
-        browserName: 'chrome',
-        javascriptEnabled: true,
-        acceptSslCerts: true,
-        chromeOptions: {
-          args: []
-        }
-      }
-    },
-
-    globals: {
-      productListUrl: `http://localhost:${PRODUCT_PORT}/productlist.html`
     }
   },
   __extend: {}
@@ -247,7 +218,6 @@ if (config.__extend.html_report_folder) {
 }
 
 const defaultOpts = config.test_settings.default.desiredCapabilities.chromeOptions;
-const chromeOpts = config.test_settings.chrome.desiredCapabilities.chromeOptions;
 
 // + 浏览器参数配置
 // headless
@@ -269,7 +239,7 @@ if (typeof config.__extend.headless === 'boolean') {
       print.log.success(`${fn.fillSpace('headless')}: ${chalk.yellow(`${config.__extend.headless}`)}`);
     }
   }
-  [defaultOpts, chromeOpts].forEach((opt) => {
+  [defaultOpts].forEach((opt) => {
     if (typeof config.__extend.headless === 'boolean') {
       // headless
       if (config.__extend.headless) {
@@ -295,7 +265,7 @@ if (config.__extend.proxy) {
   if (!iEnv.silent) {
     print.log.success(`${fn.fillSpace('proxy')}: ${chalk.yellow(`${extOs.LOCAL_IP}:${config.__extend.proxy}`)}`);
   }
-  [defaultOpts, chromeOpts].forEach((opt) => {
+  [defaultOpts].forEach((opt) => {
     opt.args.push(`--proxy-server=http=${extOs.LOCAL_IP}:${config.__extend.proxy}`);
   });
 }
@@ -305,7 +275,7 @@ if (config.__extend.userAgent) {
   if (!iEnv.silent) {
     print.log.success(`${fn.fillSpace('userAgent')}: ${chalk.yellow(config.__extend.userAgent)}`);
   }
-  [defaultOpts, chromeOpts].forEach((opt) => {
+  [defaultOpts].forEach((opt) => {
     opt.args.push(`--user-agent=${config.__extend.userAgent}`);
   });
 }